benchmark/requirements.txt
altair[all]==5.5.0
anthropic==0.60.0
gaveta==0.6.0
google-cloud-storage==3.2.0
google-genai==1.27.0
huggingface-hub[cli]==0.33.4
humanize==4.12.3
langchain-core==0.3.74
lap==0.5.12
loguru==0.7.3
more-itertools==10.7.0
mypy[faster-cache]==1.16.1
openai==1.97.0
pandas-stubs==2.3.0.250703
pandas[parquet]==2.3.1
praicing==0.0.3
python-dotenv==1.1.1
rich==14.1.0
ruff==0.11.2
scikit-learn==1.7.1
shapely==2.1.1
statsmodels==0.14.5
torchmetrics[detection]==1.8.1
transformers[torch,torch-vision,num2words,sentencepiece]==4.53.2
types-shapely==2.1.0.20250512
